:package: Refactor config module
